requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cycler==0.10.0
Django==3.1.4
dnspython==2.1.0
email-validator==1.1.2
eth-hash==0.2.0
eth-typing==2.2.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gcloud==0.18.3
googleapis-common-protos==1.52.0
greenlet==1.0.0
gunicorn==20.1.0
importlib-metadata==4.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
jws==0.1.3
kiwisolver==1.3.1
oauth2client==4.1.2
pandas==1.1.5
parsimonious==0.8.1
Pillow==8.0.1
protobuf==3.14.0
psycopg2
psycopg2-binary
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
pyparsing==2.4.7
PyQRCode==1.2.1
Pyrebase4==4.4.1
python-dotenv==0.17.0
python-jwt==2.0.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.6
scipy==1.5.4
SQLAlchemy==1.4.9
sqlparse==0.4.1
threadpoolctl==2.1.0
toolz==0.11.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1